Edited to match Pylint standards
One warning was for too many ancestors, which means that a class is inheriting from a class that itself has more than a certain number of parent classes. There isn't anything I can do about this as I need to use that class, so I set it to ignore it here.